Split tests into directories
