+++ tests/test_decomp/__init__.py
@@ -1,0 +1,1 @@
+"""Decomposition tests."""
