+++ tests/test_decomp/test_lanczos_tridiagonal.py
@@ -109,53 +109,3 @@ def _sym_tridiagonal_dense(d, e):
     offdiag1 = linalg.diagonal_matrix(e, 1)
     offdiag2 = linalg.diagonal_matrix(e, -1)
     return diag + offdiag1 + offdiag2
-
-
-@testing.parametrize("n", [50])
-@testing.parametrize("num_significant_eigvals", [4])
-@testing.parametrize("order", [6])  # ~1.5 * num_significant_eigvals
-def test_golub_kahan_lanczos_bidiagonal(A, order):
-    """Test that Lanczos tridiagonalisation yields an orthogonal-tridiagonal decomp."""
-    n, _ = np.shape(A)
-    key = prng.prng_key(1)
-    v0 = prng.normal(key, shape=(n,))
-    alg = decomp.golub_kahan_lanczos_bidiagonal(order)
-
-    def Av(v):
-        return A @ v
-
-    def vA(v):
-        return v @ A
-
-    Us, Bs, Vs, (b, v) = decomp.decompose_fori_loop(0, order + 1, v0, Av, vA, alg=alg)
-    (d_m, e_m) = Bs
-
-    tols_decomp = {"atol": 1e-5, "rtol": 1e-5}
-
-    assert np.shape(Us) == (order + 1, n)
-    assert np.allclose(Us @ Us.T, np.eye(order + 1), **tols_decomp), Us @ Us.T
-
-    assert np.shape(Vs) == (order + 1, n)
-    assert np.allclose(Vs @ Vs.T, np.eye(order + 1), **tols_decomp), Vs @ Vs.T
-
-    UAVt = Us @ A @ Vs.T
-    assert np.allclose(linalg.diagonal(UAVt), d_m, **tols_decomp)
-    assert np.allclose(linalg.diagonal(UAVt, 1), e_m, **tols_decomp)
-
-    B = _bidiagonal_dense(d_m, e_m)
-    assert np.shape(B) == (order + 1, order + 1)
-    assert np.allclose(UAVt, B, **tols_decomp)
-
-    em = np.eye(order + 1)[:, -1]
-    AVt = A @ Vs.T
-    UtB = Us.T @ B
-    AtUt = A.T @ Us.T
-    VtBtb_plus_bve = Vs.T @ B.T + b * v[:, None] @ em[None, :]
-    assert np.allclose(AVt, UtB, **tols_decomp)
-    assert np.allclose(AtUt, VtBtb_plus_bve, **tols_decomp)
-
-
-def _bidiagonal_dense(d, e):
-    diag = linalg.diagonal_matrix(d)
-    offdiag = linalg.diagonal_matrix(e, 1)
-    return diag + offdiag
