+++ tests/test_hutch/__init__.py
@@ -1,0 +1,1 @@
+"""Trace estimator tests."""
