+++ tests/test_montecarlo/__init__.py
@@ -1,0 +1,1 @@
+"""Monte-Carlo tests."""
