+++ tests/test_slq/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for stochastic Lanczos quadrature."""
