+++ tests/test_slq/test_logdet.py
